+++ src/components/manage/Blocks/Maps/MapsSidebar.test.jsx
@@ -22,7 +22,6 @@ test('renders an Image Block Sidebar component', () => {
         pathname="/news"
         onChangeBlock={() => {}}
         openObjectBrowser={() => {}}
-        resetSubmitUrl={() => {}}
       />
     </Provider>,
   );